requirements.txt
transformers[torch]
datasets
torch
torchvision
pandas
scikit-learn
pillow
mlflow
python-dotenv
openai